telegram_ucp_project/requirements_ucp.txt
fastapi
uvicorn
python-telegram-bot
httpx
sqlalchemy
pydantic
python-dotenv
supabase
